feat: sync grading sort state to url
