+++ src/pages/academy/grading/subcomponents/GradingSortIcon.tsx
@@ -1,12 +1,14 @@
 import { Button, Icon, Tooltip } from '@blueprintjs/core';
 import type { CustomHeaderProps } from 'ag-grid-react';
+import { useQueryState } from 'nuqs';
 import { useEffect, useState } from 'react';
-import { useAppDispatch, useAppSelector } from 'src/commons/utils/Hooks';
-import WorkspaceActions from 'src/commons/workspace/WorkspaceActions';
 import type { ColumnFieldsKeys, SortStateProperties } from 'src/features/grading/GradingTypes';
 import { SortStates } from 'src/features/grading/GradingTypes';
-
-import { freshSortState, getNextSortState } from './GradingSubmissionsTable';
+import {
+  freshSortState,
+  getNextSortState,
+  gradingSortParser,
+} from 'src/features/grading/GradingUtils';
 
 const SORT_TOOLTIP = {
   [SortStates.ASC]: 'Ascending',
@@ -20,20 +22,19 @@ type Props = {
 
 /**
  * Grading-specific sort control injected into the generic `ColumnHeader` via its
- * `extraActions` slot. Sorting is server-side: this only writes the desired sortyar
- * into redux (`allColsSortStates`), which `Grading.tsx` reads to build the backend
- * query. Only the actively-sorted column shows a direction.
+ * `extraActions` slot. Sorting is server-side: this only writes the desired sort
+ * into the URL (nuqs `sort`), which the overviews query reads to build the backend
+ * request. Only the actively-sorted column shows a direction.
  */
 function GradingSortIcon({ headerProps }: Props) {
   const colId = headerProps.column.getColId() as ColumnFieldsKeys;
-  const dispatch = useAppDispatch();
-  const colsSortState = useAppSelector(state => state.workspaces.grading.allColsSortStates);
+  const [colsSortState, setColsSortState] = useQueryState('sort', gradingSortParser);
   const [sortState, setSortState] = useState(SortStates.NONE);
 
   useEffect(() => {
-    if (colsSortState.sortBy !== colId) {
-      setSortState(SortStates.NONE);
-    }
+    setSortState(
+      colsSortState.sortBy === colId ? colsSortState.currentState[colId] : SortStates.NONE,
+    );
   }, [colsSortState, colId]);
 
   const handleClick = () => {
@@ -41,7 +42,7 @@ function GradingSortIcon({ headerProps }: Props) {
     setSortState(next);
     const newState: SortStateProperties = { ...freshSortState };
     newState[colId] = next;
-    dispatch(WorkspaceActions.updateAllColsSortStates({ currentState: newState, sortBy: colId }));
+    setColsSortState({ currentState: newState, sortBy: colId });
   };
 
   return (
